Add has keys request
